refactor: move remaining walkthrough components
